refactoring logo
