--- src/client/components/Logo/Logo.style.css
@@ -1,5 +1,0 @@
-.logo-img {
-  width: 75%;
-  margin-top: 1em;
-  margin-bottom: 1em;
-}
