+++ src/client/components/navigation-button/navigation-button.style.css
@@ -37,7 +37,7 @@ p {
   box-shadow: inset 2px 2px 8px #c9cbd8,
     inset -3px -3px 10px rgba(255, 255, 255, 0.9);
 }
-.item:hover .icon {
+.item:hover path {
   color: #73abd7;
 }
 .item:hover p {
