+++ src/client/components/side-navigation/sidebar.stories.js
