+++ src/client/components/side-navigation/sidebar.style.css
@@ -9,3 +9,8 @@
      background-color: #ECEDF1;
  }
  
+ .sidebar-wrapper .logo{
+    margin: 1em 0;  
+ }
+
+
